Basic image uploading. Writes a new picture to a folder in the public directory, adds path, description, and hard-coded author_id (1), so make sure you register at least 1 user. Can't view uploads yet, doesn't track user, login not required yet.
For my test views, I used Blade, and some Form controls not included by default
in Laravel 5. Follow the steps at this link to get it up: http://laravel.io/forum/09-20-2014-html-form-class-not-found-in-laravel-5?page=1#reply-21628
Specifically that comment I linked to, since it summarizes the steps. If you get an error about bindShared() function, open the file it's complaining about
and change all instances (2 of them?) of bindshared() to singleton().

Be sure to copy the new version of the database from the google drive. Contact me with issues.